点のグリッドは掛け算ではなく足し算
`33 × 15.6969697` は最後のビットまで 518.0 ですが、`15.6969697` を 33 回
足すと 517.99988 です。SAMPLE1 の 34 本目の列は前者だと 639.0（作図領域の
外）、後者だと 638.99988。**本物は 638 に打ちます**——グリッドの 28 行
ぜんぶで 1 点ずつ。ほかの列と行はどちらでも同じ値になります。

刻み自体も `mm × (518/紙幅)` ではなく `mm × 518 / 紙幅` を double で持ちます。

| | 前 | 後 |
|---|---|---|
| SAMPLE1 | 70 | 42 |
| 合計 | 386 | **358** |

+++ src/jwc.h
@@ -79,7 +79,12 @@ typedef struct {
     /* The dot grid: whether it is on, and its spacing in drawing units.  It
      * runs through the drawing's origin in both directions. */
     unsigned char grid_on;
-    float grid_x, grid_y;
+    /* Kept as doubles, and worked out as `mm * 518 / paper` rather than
+     * `mm * (518 / paper)`.  The two differ in the last bit, and SAMPLE1's
+     * thirty-fourth column lands on 639.0000095 one way and 638.99999999999 the
+     * other -- one pixel off the right edge of the drawing area, so the port
+     * dropped a dot the original draws, once on each of the grid's 28 rows. */
+    double grid_x, grid_y;
     /* How many drawing units a millimetre of paper is: JW_CAD's own drawing
      * area (518 pixels) over the paper's width.  The character sizes are in
      * millimetres, so this is what turns them into pixels. */
